+++ src/Inventory/Application/Command/DevicesNotFound.php
@@ -1,0 +1,10 @@
+<?php
+
+/*
+ * SPDX-FileCopyrightText: 2026 The Kadupul project and contributors
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+namespace Kadupul\Inventory\Application\Command;
+
+final class DevicesNotFound extends \RuntimeException {}